# specs/001-nightlife-directory/contracts/api.yaml
...
openapi: 3.0.3
info:
  title: Nightnice API
  description: |
    REST API for Thailand Nightlife Directory.
    - Public endpoints: No authentication required
    - Admin endpoints: JWT Bearer token required
  version: 1.0.0
  contact:
    name: Nightnice Team
servers:
  - url: https://api.nightnice.com
    description: Production
  - url: https://api-staging.nightnice.com
    description: Staging
  - url: http://localhost:5000
    description: Local Development
...
paths:
...
  /api/stores:
    get:
      tags: [Stores]
      summary: List stores with filters
      description: Get paginated list of active stores with optional filters
      parameters:
        - name: province
          in: query
          schema:
            type: string
          description: Filter by province slug
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Search term (name, description)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page
        - name: featured
          in: query
          schema:
            type: boolean
          description: Filter featured stores only
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created_at, -name, -created_at]
            default: -created_at
          description: Sort field (prefix - for descending)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/stores/{slug}:
    get:
      tags: [Stores]
      summary: Get store by slug
      description: Get detailed store information by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/stores/{slug}/nearby:
    get:
      tags: [Stores]
      summary: Get nearby stores
      description: Get stores within specified radius of the given store
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: radius
          in: query
          schema:
            type: number
            default: 5
            minimum: 1
            maximum: 50
          description: Search radius in kilometers
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
            minimum: 1
            maximum: 20
          description: Maximum number of stores to return
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyStoresResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/provinces/{slug}:
    get:
      tags: [Provinces]
      summary: Get province by slug
      description: Get province details with store count by category
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvinceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/ads:
    get:
      tags: [Advertisements]
      summary: Get active advertisements
      description: Get ads targeted for the current page context
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [banner, sponsored, featured]
          description: Filter by ad type
        - name: province
          in: query
          schema:
            type: string
          description: Province slug for targeting
        - name: category
          in: query
          schema:
            type: string
          description: Category slug for targeting
        - name: page_context
          in: query
          schema:
            type: string
            maxLength: 100
          description: Current page context for tracking
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdListResponse'
  /api/ads/track:
    post:
      tags: [Advertisements]
      summary: Track ad event
      description: Record impression or click event for an ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdTrackRequest'
      responses:
        '204':
          description: Event recorded successfully
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/admin/stores:
    get:
      tags: [Admin Stores]
      summary: List all stores (Admin)
      description: Get all stores including inactive ones
      security:
        - BearerAuth: []
      parameters:
        - name: province
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Admin Stores]
      summary: Create store
      description: Create a new store
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreCreateRequest'
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/stores/{id}:
    get:
      tags: [Admin Stores]
      summary: Get store by ID (Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Admin Stores]
      summary: Update store
      description: Update an existing store
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUpdateRequest'
      responses:
        '200':
          description: Store updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Admin Stores]
      summary: Delete store
      description: Permanently delete a store
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Store deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/provinces:
    get:
      tags: [Admin Provinces]
      summary: List all provinces (Admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProvinceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/provinces/{id}:
    put:
      tags: [Admin Provinces]
      summary: Update province
      description: Update province SEO description and sort order
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvinceUpdateRequest'
      responses:
        '200':
          description: Province updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProvinceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # ADMIN CATEGORY MANAGEMENT
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error
        - message
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - limit
        - total
        - total_pages
...
    StoreListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logo_url:
          type: string
          format: uri
          nullable: true
        province:
          $ref: '#/components/schemas/ProvinceRef'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryRef'
        price_range:
          type: integer
          minimum: 1
          maximum: 4
          nullable: true
        is_featured:
          type: boolean
      required:
        - id
        - name
        - slug
        - province
        - categories
        - is_featured
    StoreListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StoreListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required:
        - data
        - meta
    StoreDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        logo_url:
          type: string
          format: uri
          nullable: true
        banner_url:
          type: string
          format: uri
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        google_map_url:
          type: string
          format: uri
          nullable: true
        line_id:
          type: string
          nullable: true
        facebook_url:
          type: string
          format: uri
          nullable: true
        instagram_url:
          type: string
          format: uri
          nullable: true
        price_range:
          type: integer
          minimum: 1
          maximum: 4
          nullable: true
        open_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        close_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        facilities:
          type: array
          items:
            type: string
            enum: [parking, live_music, karaoke, outdoor_seating, wifi, reservation]
        is_featured:
          type: boolean
        province:
          $ref: '#/components/schemas/ProvinceRef'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryRef'
        gallery:
          type: array
          items:
            $ref: '#/components/schemas/StoreImageRef'
      required:
        - id
        - name
        - slug
        - facilities
        - is_featured
        - province
        - categories
        - gallery
    StoreImageRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        sort_order:
          type: integer
      required:
        - id
        - url
        - sort_order
    NearbyStoresResponse:
      type: object
      properties:
        data:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/StoreListItem'
              - type: object
                properties:
                  distance_km:
                    type: number
                    format: double
      required:
        - data
    StoreCreateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        province_id:
          type: string
          format: uuid
        category_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        description:
          type: string
          nullable: true
        phone:
          type: string
          maxLength: 20
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          nullable: true
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          nullable: true
        google_map_url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        line_id:
          type: string
          maxLength: 100
          nullable: true
        facebook_url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        instagram_url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        price_range:
          type: integer
          minimum: 1
          maximum: 4
          nullable: true
        open_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        close_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        facilities:
          type: array
          items:
            type: string
            enum: [parking, live_music, karaoke, outdoor_seating, wifi, reservation]
        is_active:
          type: boolean
          default: true
        is_featured:
          type: boolean
          default: false
      required:
        - name
        - province_id
        - category_ids
    StoreUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        province_id:
          type: string
          format: uuid
        category_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        description:
          type: string
          nullable: true
        phone:
          type: string
          maxLength: 20
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          nullable: true
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          nullable: true
        google_map_url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        line_id:
          type: string
          maxLength: 100
          nullable: true
        facebook_url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        instagram_url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        price_range:
          type: integer
          minimum: 1
          maximum: 4
          nullable: true
        open_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        close_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        facilities:
          type: array
          items:
            type: string
            enum: [parking, live_music, karaoke, outdoor_seating, wifi, reservation]
        is_active:
          type: boolean
        is_featured:
          type: boolean
    AdminStoreResponse:
      allOf:
        - $ref: '#/components/schemas/StoreDetailResponse'
        - type: object
          properties:
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    AdminStoreListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/StoreListItem'
              - type: object
                properties:
                  is_active:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required:
        - data
        - meta

    # ============================================
    # PROVINCE SCHEMAS
    # ============================================
    ProvinceRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
...
    ProvinceDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        seo_description:
          type: string
          nullable: true
        region:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
          required:
            - id
            - name
            - slug
        category_counts:
          type: array
          items:
            type: object
            properties:
              category:
                $ref: '#/components/schemas/CategoryRef'
              count:
                type: integer
      required:
        - id
        - name
        - slug
        - region
        - category_counts
    ProvinceUpdateRequest:
      type: object
      properties:
        seo_description:
          type: string
          nullable: true
        sort_order:
          type: integer
    AdminProvinceResponse:
      allOf:
        - $ref: '#/components/schemas/ProvinceDetailResponse'
        - type: object
          properties:
            sort_order:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    AdminProvinceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminProvinceResponse'
      required:
        - data
...
    CategoryRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
...
    AdListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [banner, sponsored, featured]
        image_url:
          type: string
          format: uri
          nullable: true
        target_url:
          type: string
          format: uri
          nullable: true
        store:
          $ref: '#/components/schemas/StoreListItem'
          nullable: true
      required:
        - id
        - type
    AdListResponse:
      type: object
      properties:
        banners:
          type: array
          items:
            $ref: '#/components/schemas/AdListItem'
        sponsored:
          type: array
          items:
            $ref: '#/components/schemas/AdListItem'
        featured:
          type: array
          items:
            $ref: '#/components/schemas/AdListItem'
      required:
        - banners
        - sponsored
        - featured
    AdTrackRequest:
      type: object
      properties:
        ad_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [impression, click]
        page_context:
          type: string
          maxLength: 100
      required:
        - ad_id
        - event_type
...